requirements.txt
alembic==1.0.5
certifi==2018.11.29
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
psycopg2==2.7.6.1
python-dateutil==2.7.5
python-editor==1.0.3
six==1.12.0
SQLAlchemy==1.2.15
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
tensorflow==2.0.0
pillow
numpy
pybase64